feat: Enhance ESLint configuration; disable unused vars and explicit any rules feat: Refactor ForgetPassword component; replace var with const for better readability feat: Update LastStatement and CustomerDetails components; adjust import paths for column definitions feat: Move column definitions to separate files for Invoice and TransactionList components feat: Improve DashboardMainLayout; add hasAccess to dependency array for useEffect feat: Add search functionality to Users page; enhance user experience feat: Update NotFound component; replace img with Image component for better performance feat: Refactor AppHeader component; replace var with const for password handling feat: Optimize Search component; adjust dependencies for useCallback feat: Update next.config.mjs; ensure proper image domain configuration
